feat(1061): rename files, amend imports
