+++ packages/forms-web-app/src/routes/rule-6-appeals/index.js
@@ -1,8 +1,8 @@
 const express = require('express');
 const router = express.Router();
 
-const selectedAppealController = require('../../controllers/selected-appeal/selected-appeal');
-const appealDetailsController = require('../../controllers/selected-appeal/appeal-details/appeal-details');
+const selectedAppealController = require('../../controllers/selected-appeal');
+const appealDetailsController = require('../../controllers/selected-appeal/appeal-details');
 
 router.get('/:appealNumber', selectedAppealController.get);
 router.get('/:appealNumber/appeal-details', appealDetailsController.get);
